+++ src/main/java/domain/deck/Deck.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.deck;
 
 import domain.card.Card;
