+++ src/test/java/domain/BlackJackGameTest.java
@@ -3,10 +3,11 @@ package domain;
 import domain.card.Card;
 import domain.card.Rank;
 import domain.card.Symbol;
+import domain.deck.Deck;
 import domain.gamer.Dealer;
 import domain.gamer.Name;
 import domain.gamer.Player;
-import domain.strategy.SettedDeckGenerator;
+import domain.deck.strategy.SettedDeckGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
